Move files from database/utilities to database to have accessible embedded mgirations folder
