+++ backend/internal/database/checkIfFileIsInBatchByPath.go
