+++ backend/internal/database/getDistinctFiles.go
