--- backend/internal/database/utilities/getSQLFileContent.go
@@ -1,18 +1,0 @@
-package utilities
-
-import (
-	"os"
-	"path"
-)
-
-// TODO CHECK PATHS ON EXECUTION
-func GetSQLFileContent(filename string) (string, error) {
-	var relativeFilePathToMigrationFolder = path.Join("./migrations/", filename)
-	contentAsByteArray, readFileError := os.ReadFile(relativeFilePathToMigrationFolder)
-	if readFileError != nil {
-		return "", readFileError
-	}
-
-	var stringifiedContent = string(contentAsByteArray)
-	return stringifiedContent, nil
-}
